Remove Unused URIDryConverter.java
